% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summernote.R
\name{summernoteInput}
\alias{summernoteInput}
\title{Include a summernote editor in the UI}
\usage{
summernoteInput(inputId, label = NULL, value = "", ...)
}
\arguments{
\item{inputId}{ID of the corresponding input}

\item{label}{Label of the input}

\item{value}{Initial value of the editor}

\item{...}{Other arguments used to initialize the editor}
}
\description{
This functions create the neccessary HTML-tags to include
summernote wysiwyg editor within the apps UI.
}
