% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summernote.R
\name{updateSummernoteInput}
\alias{updateSummernoteInput}
\title{Update summernote editor}
\usage{
updateSummernoteInput(inputId, value,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{ID of the summernote editor}

\item{value}{New content of the editor}
}
\description{
This function updates the content of summernote editor
with the given ID.
}
\seealso{
\code{\link{summernoteInput}}
}
